requirements.txt
streamlit==1.31.1
altair==4.2.2
pandas
numpy
plotly
python-dotenv
requests
scikit-learn
psycopg[binary]>=3.2
fastapi>=0.115
uvicorn[standard]>=0.30
